+++ src/console/menu/camera_menu.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   camera_menu.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: shurtado <shurtado@student.42.fr>          +#+  +:+       +#+        */
+/*   By: shurtado <shurtado@student.42barcelona.fr> +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/17 11:18:43 by shurtado          #+#    #+#             */
-/*   Updated: 2024/12/17 11:19:05 by shurtado         ###   ########.fr       */
+/*   Updated: 2025/01/07 12:41:36 by shurtado         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "miniRT.h"
 
-void	adjust_camera_pos(keys_t key, t_data *data, float *posq)
+void	adjust_camera_pos(keys_t key, t_data *data, double *posq)
 {
 	if (key == MLX_KEY_UP)
 		data->cam->pos.y += *posq;
@@ -27,37 +27,37 @@ void	adjust_camera_pos(keys_t key, t_data *data, float *posq)
 	else if (key == MLX_KEY_X)
 		data->cam->pos.z += *posq;
 	else if (key == MLX_KEY_MINUS || key == MLX_KEY_KP_SUBTRACT)
-		*posq = fmaxf(1.0f, *posq - 1);
+		*posq = fmax(1.0f, *posq - 1);
 	else if (key == MLX_KEY_KP_ADD || key == MLX_KEY_EQUAL)
-		*posq = fminf(100, *posq + 1);
+		*posq = fmin(100, *posq + 1);
 }
 
-void	adjust_camera_axis(keys_t key, t_data *data, float *axisq)
+void	adjust_camera_axis(keys_t key, t_data *data, double *axisq)
 {
 	if (key == MLX_KEY_UP)
-		data->cam->axis.y = fminf(data->cam->axis.y + *axisq, CAMAXISMAX);
+		data->cam->axis.y = fmin(data->cam->axis.y + *axisq, CAMAXISMAX);
 	else if (key == MLX_KEY_DOWN)
-		data->cam->axis.y = fmaxf(data->cam->axis.y - *axisq, CAMAXISMIN);
+		data->cam->axis.y = fmax(data->cam->axis.y - *axisq, CAMAXISMIN);
 	else if (key == MLX_KEY_LEFT)
-		data->cam->axis.x = fmaxf(data->cam->axis.x - *axisq, CAMAXISMIN);
+		data->cam->axis.x = fmax(data->cam->axis.x - *axisq, CAMAXISMIN);
 	else if (key == MLX_KEY_RIGHT)
-		data->cam->axis.x = fminf(data->cam->axis.x + *axisq, CAMAXISMAX);
+		data->cam->axis.x = fmin(data->cam->axis.x + *axisq, CAMAXISMAX);
 	else if (key == MLX_KEY_Z)
-		data->cam->axis.z = fmaxf(data->cam->axis.z - *axisq, CAMAXISMIN);
+		data->cam->axis.z = fmax(data->cam->axis.z - *axisq, CAMAXISMIN);
 	else if (key == MLX_KEY_X)
-		data->cam->axis.z = fminf(data->cam->axis.z + *axisq, CAMAXISMAX);
+		data->cam->axis.z = fmin(data->cam->axis.z + *axisq, CAMAXISMAX);
 	else if (key == MLX_KEY_MINUS || key == MLX_KEY_KP_SUBTRACT)
-		*axisq = fmaxf(0.01f, *axisq - 0.01f);
+		*axisq = fmax(0.01f, *axisq - 0.01f);
 	else if (key == MLX_KEY_KP_ADD || key == MLX_KEY_EQUAL)
-		*axisq = fminf(1.0f, *axisq + 0.01f);
+		*axisq = fmin(1.0f, *axisq + 0.01f);
 	data->cam->axis = normalize(data->cam->axis);
 }
 
 void	manage_cam(keys_t key, t_data *data)
 {
 	static bool		mode;
-	static float	posq;
-	static float	axisq;
+	static double	posq;
+	static double	axisq;
 
 	if (posq < 1.0f)
 		posq = CAMPLUS;
